refactor: split database services
